Splitting up combinators.py
