+++ merlin/models/tf/features/continuous.py
@@ -19,7 +19,7 @@ from typing import List, Optional
 import tensorflow as tf
 
 from merlin.models.tf.blocks.core.base import BlockType
-from merlin.models.tf.blocks.core.combinators import (
+from merlin.models.tf.blocks.core.tabular import (
     TABULAR_MODULE_PARAMS_DOCSTRING,
     Filter,
     TabularAggregationType,
